Move forbidden error to a leaf module; respect the access-service ratchet
access-service.ts is debt-tracked (#1520) and must not grow, so
EngramAccessForbiddenError lives in a new access-errors.ts leaf module
(same shape as abort-error.ts). Behavior unchanged.
