+++ packages/remnic-core/src/access-errors.ts
@@ -1,0 +1,2 @@
+/** Authenticated caller lacks the required authorization (HTTP 403). */
+export class EngramAccessForbiddenError extends Error {}
