+++ packages/remnic-core/src/access-service.ts
@@ -260,9 +260,6 @@ import { selfDeps } from "./orchestration/self-deps.js";
 
 export class EngramAccessInputError extends Error {}
 
-/** Authenticated caller lacks the required authorization (HTTP 403). */
-export class EngramAccessForbiddenError extends Error {}
-
 function qmdCollectionPathParts(resultPath: string): {
   collection: string;
   relativePath: string;
